machine_learning_devops_engineer/deploying-a-scalable-pipeline-in-prod-demos/project-deploy-ML-model-FastAPI/requirements.txt
#python==3.11.5
flake8
pytest==7.4.2
pandas==2.0.3
scikit-learn==1.3.0
pydantic==2.0.3
pydantic-core==2.3.0
pyyaml==6.0.1
fastapi==0.103.1
httpx==0.25.0 
wandb==0.15.10
numpy==1.23.5
requests==2.31.0 
uvicorn==0.23.2
gunicorn==21.2.0
